+++ buylocalServer/server/models/bewertung.js
@@ -1,10 +1,11 @@
 
 const Sequelize = require('sequelize');
+const moment = require('moment');
 const sequelize = new Sequelize("buylocal","buylocalAPI","buyl0cal",{host: 'localhost', dialect:"mysql",logging:false});
 
  const Bewertung= sequelize.define('Bewertung',{
     BewertungID: {type: Sequelize.INTEGER, primaryKey:true, autoIncrement:true},
-    Datum: Sequelize.DATE,
+    Datum: {type:Sequelize.DATE,    get: function() {return moment.utc(this.getDataValue('last_login')).format('HH:MM DD.MM.YYYY')}},
     Sterne: Sequelize.DOUBLE,
     Bewerter: Sequelize.INTEGER,
     Bewerteter:Sequelize.INTEGER,
